Wrap onboarding page in suspense to fix SSR issue; create onboarding form component
